in_thermal: rename temp input plugin to thermal
