+++ CMakeLists.txt
@@ -94,7 +94,7 @@ option(FLB_PROXY_GO           "Enable Go plugins support"   Yes)
 
 # Build-in Plugins
 option(FLB_IN_CPU             "Enable CPU input plugin"            Yes)
-option(FLB_IN_TEMP            "Enable Temperature plugin"          Yes)
+option(FLB_IN_THERMAL         "Enable Thermal plugin"              Yes)
 option(FLB_IN_DISK            "Enable Disk input plugin"           Yes)
 option(FLB_IN_EXEC            "Enable Exec input plugin"           Yes)
 option(FLB_IN_FORWARD         "Enable Forward input plugin"        Yes)
