+++ plugins/CMakeLists.txt
@@ -91,7 +91,7 @@ endmacro()
 if(${CMAKE_SYSTEM_NAME} MATCHES "Linux")
   REGISTER_IN_PLUGIN("in_cpu")
   REGISTER_IN_PLUGIN("in_mem")
-  REGISTER_IN_PLUGIN("in_temp")
+  REGISTER_IN_PLUGIN("in_thermal")
   REGISTER_IN_PLUGIN("in_kmsg")
   REGISTER_IN_PLUGIN("in_proc")
   REGISTER_IN_PLUGIN("in_disk")
