--- plugins/in_temp/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(src
-  in_temp.c)
-
-FLB_PLUGIN(in_temp "${src}" "")
