+++ plugins/in_thermal/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(src
+  in_thermal.c)
+
+FLB_PLUGIN(in_thermal "${src}" "")
