+++ plugins/in_thermal/in_thermal.h
